rename command and add to based
