+++ pkg/cmd/p2p.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"net/http"
 	"strings"
+	"text/tabwriter"
 
 	"connectrpc.com/connect"
 	"github.com/spf13/cobra"
@@ -14,8 +15,8 @@ import (
 )
 
 // NodeInfoCmd returns information about the running node via RPC
-var NodeInfoCmd = &cobra.Command{
-	Use:   "node-info",
+var NetInfoCmd = &cobra.Command{
+	Use:   "net-info",
 	Short: "Get information about a running node via RPC",
 	Long:  "This command retrieves the node information via RPC from a running node in the specified directory (or current directory if not specified).",
 	RunE: func(cmd *cobra.Command, args []string) error {
@@ -52,24 +53,26 @@ var NodeInfoCmd = &cobra.Command{
 			return fmt.Errorf("error calling GetNetInfo RPC: %w", err)
 		}
 
-		// Print node information with better formatting
 		netInfo := resp.Msg.NetInfo
-		nodeID := netInfo.Id // Store Node ID separately
+		nodeID := netInfo.Id
 
-		fmt.Println("\n" + strings.Repeat("=", 50))
-		fmt.Println("📊 NODE INFORMATION")
-		fmt.Println(strings.Repeat("=", 50))
-		fmt.Printf("🆔 Node ID:      \033[1;36m%s\033[0m\n", nodeID) // Print Node ID once
+		out := cmd.OutOrStdout()
+		w := tabwriter.NewWriter(out, 2, 0, 2, ' ', 0)
+
+		fmt.Fprintf(w, "%s", strings.Repeat("=", 50))
+		fmt.Fprintf(w, "📊 NODE INFORMATION")
+		fmt.Fprintf(w, "%s\n", strings.Repeat("=", 50))
+		fmt.Fprintf(w, "🆔 Node ID:      \033[1;36m%s\033[0m\n", nodeID) // Print Node ID once
 
 		// Iterate through all listen addresses
-		fmt.Println("📡 Listen Addrs:")
+		fmt.Fprintf(w, "📡 Listen Addrs:")
 		for i, addr := range netInfo.ListenAddresses {
 			fullAddress := fmt.Sprintf("%s/p2p/%s", addr, nodeID)
-			fmt.Printf("   [%d] Addr: \033[1;36m%s\033[0m\n", i+1, addr)
-			fmt.Printf("       Full: \033[1;32m%s\033[0m\n", fullAddress)
+			fmt.Fprintf(w, "   [%d] Addr: \033[1;36m%s\033[0m\n", i+1, addr)
+			fmt.Fprintf(w, "       Full: \033[1;32m%s\033[0m\n", fullAddress)
 		}
 
-		fmt.Println(strings.Repeat("-", 50))
+		fmt.Fprintf(w, "%s\n", strings.Repeat("-", 50))
 		// Also get peer information
 		peerResp, err := p2pClient.GetPeerInfo(
 			context.Background(),
@@ -81,12 +84,12 @@ var NodeInfoCmd = &cobra.Command{
 
 		// Print connected peers in a table-like format
 		peerCount := len(peerResp.Msg.Peers)
-		fmt.Printf("👥 CONNECTED PEERS: \033[1;33m%d\033[0m\n", peerCount)
+		fmt.Fprintf(w, "👥 CONNECTED PEERS: \033[1;33m%d\033[0m\n", peerCount)
 
 		if peerCount > 0 {
-			fmt.Println(strings.Repeat("-", 50))
-			fmt.Printf("%-5s %-20s %s\n", "NO.", "PEER ID", "ADDRESS")
-			fmt.Println(strings.Repeat("-", 50))
+			fmt.Fprintf(w, "%s\n", strings.Repeat("-", 50))
+			fmt.Fprintf(w, "%-5s %-20s %s\n", "NO.", "PEER ID", "ADDRESS")
+			fmt.Fprintf(w, "%s\n", strings.Repeat("-", 50))
 
 			for i, peer := range peerResp.Msg.Peers {
 				// Truncate peer ID if it's too long for display
@@ -94,13 +97,13 @@ var NodeInfoCmd = &cobra.Command{
 				if len(peerID) > 18 {
 					peerID = peerID[:15] + "..."
 				}
-				fmt.Printf("%-5d \033[1;34m%-20s\033[0m %s\n", i+1, peerID, peer.Address)
+				fmt.Fprintf(w, "%-5d \033[1;34m%-20s\033[0m %s\n", i+1, peerID, peer.Address)
 			}
 		} else {
-			fmt.Println("\n\033[3;33mNo peers connected\033[0m")
+			fmt.Fprintf(w, "\n\033[3;33mNo peers connected\033[0m")
 		}
 
-		fmt.Println(strings.Repeat("=", 50) + "\n")
+		fmt.Fprintf(w, "%s\n", strings.Repeat("=", 50))
 
 		return nil
 	},
